+++ TheWorld/src/TheWorld/Models/Stop.cs
@@ -9,8 +9,8 @@ namespace TheWorld.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public long Longitude { get; set; }
-        public long Latitude { get; set; }
+        public double Longitude { get; set; }
+        public double Latitude { get; set; }
         public DateTime Arrival { get; set; }
         public int Order { get; set; }
     }
